Split out pipermail to a main file vs pipermail specific functions. Verfied runs.
